1. extracted the TaskErrorBoundary from the TaskList component and importing it instead, to reduce clutter 2. placed task-list and task-item files in there own separate directories, and created an error-handling directory to place any other error boundaries or error related components
