+++ src/components/error-handling/task-error-boundary.tsx
@@ -1,0 +1,31 @@
+import { Component } from "react";
+
+export default class TaskErrorBoundary extends Component<
+  { children: React.ReactNode },
+  { hasError: boolean }
+> {
+  constructor(props: { children: React.ReactNode }) {
+    super(props);
+    this.state = { hasError: false };
+  }
+
+  static getDerivedStateFromError() {
+    return { hasError: true };
+  }
+
+  componentDidCatch(error: Error, errorInfo: React.ErrorInfo): void {
+    console.error(error, errorInfo);
+  }
+
+  render() {
+    if (this.state.hasError) {
+      return (
+        <div className="text-red-600 p-4 text-center">
+          Something went wrong. Please try again.
+        </div>
+      );
+    }
+
+    return this.props.children;
+  }
+}
