+++ src/components/task-list/task-list.tsx
@@ -2,7 +2,8 @@
 
 import { useEffect, useState, Component } from "react";
 import { Task } from "@/types/task";
-import TaskItem from "./task-item";
+import TaskItem from "../task-item/task-item";
+import TaskErrorBoundary from "../error-handling/task-error-boundary";
 import React from "react";
 
 const LoadingState = () => (
@@ -16,45 +17,14 @@ const LoadingState = () => (
   </div>
 );
 
-class TaskErrorBoundary extends Component<
-  { children: React.ReactNode },
-  { hasError: boolean }
-> {
-  constructor(props: { children: React.ReactNode }) {
-    super(props);
-    this.state = { hasError: false };
-  }
-
-  static getDerivedStateFromError() {
-    return { hasError: true };
-  }
-
-  componentDidCatch(error: Error, errorInfo: React.ErrorInfo): void {
-    console.error(error, errorInfo);
-  }
-
-  render() {
-    if (this.state.hasError) {
-      return (
-        <div className="text-red-600 p-4 text-center">
-          Something went wrong. Please try again.
-        </div>
-      );
-    }
-
-    return this.props.children;
-  }
-}
-
+// todo: there are a lot of useState calls lined up, this should be reduced and customized
 export default function TaskList() {
-  // use a custom hook for managing the state, 
   const [tasks, setTasks] = useState<Task[]>([]);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
   const [page, setPage] = useState(1);
   const [hasMore, setHasMore] = useState(true);
 
-  // memoize fetchTasks and only recreate if page has changed
   const fetchTasks = React.useCallback(async (pageNum: number) => {
     try {
       setLoading(true);
@@ -90,7 +60,7 @@ export default function TaskList() {
           <TaskItem key={task.id} task={task} />
         ))}
         {loading && <LoadingState />}
-        {loading && !hasMore &&
+        {!loading && hasMore &&
           <button
             onClick={handleLoadMore}
             className="w-full p-4 text-center font-bold text-gray-600 hover:text-gray-900"
